Working decryption and proof checks

+++ src/state.rs
@@ -5,7 +5,7 @@ use cosmwasm_schema::{cw_serde};
 use crate::msg::{Poll};
 
 pub const ENCRYPTED_VOTE: Item<Vec<Vec<String>>> = Item::new("encrypted_vote");
-pub const DECRYPTED_TALLY: Item<Vec<i32>> = Item::new("tally");
+pub const DECRYPTED_TALLY: Item<Vec<u64>> = Item::new("tally");
 pub const POLL: Item<Poll> = Item::new("poll");
 pub const TOTAL_VOTES: Item<u32> = Item::new("total_votes");
 pub const POLL_PUB_KEY: Item<String> = Item::new("poll_pub_key");
